Modify: MentionList, ChoiceList 분리
