+++ src/components/Home/MentionInput.tsx
@@ -1,11 +1,10 @@
 import { FormEvent, useRef, useState, KeyboardEvent, useEffect } from "react";
 
 import { Input } from "@/components/ui/input.tsx";
-import { Badge } from "@/components/ui/badge.tsx";
-import { Separator } from "@/components/ui/separator.tsx";
 
 import trie, { MyType } from "@/lib/trie.ts";
-import { cn } from "@/lib/utils.ts";
+import MentionList from "@/components/Home/MentionList.tsx";
+import ChoiceList from "@/components/Home/ChoiceList.tsx";
 
 const MentionInput = () => {
   const [mentionList, setMentionList] = useState<Array<MyType>>([]);
@@ -20,12 +19,15 @@ const MentionInput = () => {
     list && setMentionList(list);
   };
 
-  const handleAddChoiceList = (people: MyType) => {
+  const initInput = () => {
     if (!inputRef.current) return;
-    inputRef.current.value = "";
-    setChoiceList((prev) => [...prev, people]);
     setMentionList([]);
     inputRef.current.value = "";
+  };
+
+  const handleAddChoiceList = (people: MyType) => {
+    setChoiceList((prev) => [...prev, people]);
+    initInput();
   };
 
   const handleDeleteChoiceList = (people: MyType) => {
@@ -53,8 +55,7 @@ const MentionInput = () => {
           ...prev,
           { name: mentionList[focusIdx].name, userId: mentionList[focusIdx].userId },
         ]);
-        setMentionList([]);
-        inputRef.current.value = "";
+        initInput();
         break;
     }
   };
@@ -83,52 +84,3 @@ const MentionInput = () => {
 };
 
 export default MentionInput;
-
-interface ListProps {
-  list: MyType[];
-  onClick: (people: MyType) => void;
-}
-
-interface MentionListProps extends ListProps {
-  focusIdx: number;
-}
-
-const MentionList = ({ list, onClick, focusIdx }: MentionListProps) => {
-  return (
-    <div className="mt-2 overflow-hidden scroll-auto border p-2">
-      {list.map(({ name, userId }, idx) => {
-        return (
-          <div
-            key={userId}
-            className={cn(
-              "hover:bg-gray-100 focus:bg-gray-100",
-              focusIdx === idx ? "bg-gray-100" : "",
-            )}
-          >
-            <p
-              onClick={() => onClick({ name, userId })}
-              className="cursor-pointer py-2 hover:font-bold focus:font-bold"
-            >
-              {name}
-            </p>
-            {idx !== list.length - 1 && <Separator />}
-          </div>
-        );
-      })}
-    </div>
-  );
-};
-
-const ChoiceList = ({ list, onClick }: ListProps) => {
-  return (
-    <div className="mb-2 flex gap-2">
-      {list.map(({ name, userId }) => {
-        return (
-          <Badge key={userId} onClick={() => onClick({ name, userId })} className="cursor-pointer">
-            {name}
-          </Badge>
-        );
-      })}
-    </div>
-  );
-};
